+++ src/multiplex/server.rs
@@ -1,4 +1,5 @@
-use super::{multiplex, RequestId, Error, Message, ServerService, Transport};
+use {Error};
+use super::{multiplex, RequestId, Message, ServerService, Transport};
 use futures::{Future, Poll, Async};
 use std::io;
